updated jenkins scripts with restart function

+++ spring-boot-default-skeleton/jenkins/jenkins-redeploy.sh
@@ -6,7 +6,7 @@ TAG=latest
 SPRING_BOOT_PROFILE=prod
 
 if [ "$(docker ps -q -f name=${APP})" ]; then docker kill ${APP}; else echo "Container ${APP} is missing"; fi
-docker rm ${APP}
+if [ "$(docker ps -q -a -f name=${APP})" ]; then docker rm ${APP}; else echo "Container ${APP} is missed."; fi
 
 docker pull ${COMPANY}/${APP}:${TAG}
