+++ spring-boot-default-skeleton/jenkins/jenkins-stop.sh
@@ -3,4 +3,4 @@
 APP=spring-boot-server
 
 if [ "$(docker ps -q -f name=${APP})" ]; then docker kill ${APP}; else echo "Container ${APP} is missing"; fi
-docker rm ${APP}
+if [ "$(docker ps -q -a -f name=${APP})" ]; then docker rm ${APP}; else echo "Container ${APP} is missed."; fi
